requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
python-multipart==0.0.6
streamlit==1.29.0

